Create License
This HTML file "otherhtm.htm" was composed from examples on a HTML learning web site and customized to provide access to stored information on the Othernet Dream catcher boards SD secondary memory card. This software is placed in the public domain and I encourage anyone with HTML skills to improve upon it.